Galapagos now takes variable PSF
The scripts have been changed to support variable PSF now.
Leaving verions it as it's own branch for now.
More testing needs to be done, possibly on real GAMA data (non-convolved images),
I have asked Lee how he creates his PSFs and would run that on lots of positions to
create a long list of psfs for galapagos.

Had to change galapagos.pro and gala_bridge.pro slightly, but main changes are 2
new scripts, readin_psf_file.pro and choose_psf.pro.
Both are sufficiently easy and can be extended for further flexibility if needed/wanted.

The scripts support the following input options now:
in the galapagos setup file, give the file in the PSF section
1) a single psf file.
The given PSF file will be chosen for the fit of every object.
The input file should be a fits image, including the '.fits' at the end (this is how it notices!)
2) a psf per tile.
Galapagos runs on big tiles (e.g. cut up GAMA data or HST shots within GEMS/STAGES).
When this format is (automatically) detected, the script checks whether a PSF is given for every tile.
If not, it prints a warning and stops!
The input file should be an ascii list with
tilename psfname (separated by ' ', one line per tile/psf)
3) on a 'closest' basis.
the CLOSEST PSF to the objects position will be chosen.
the input file should be
RA DEC PSF (separated by ' ', one line per tile/psf)
4) on a 'box' setting.
define boxes of RA & DEC in which a certain PSF should be used
input file should be
RA_min RA_max DEC_min DEC_max PSF (separated by ' ', one line per tile/psf)
The skript first checks whether there are objects that lie in several (or no) of the defined boxes,
and prints out a warning (but keeps running)
For objects within 1 (!) box, the according PSF is chosen
For objects within 0 boxes, the closest PSF is chosen (as defined by the centers of the defined boxes!!)
If the object lies within SEVERAL of the defined boxes, the closer of THESE (as given by box centers)
is chosen. This way, each object WILL get a PSF!

I further changed galapagos.pro to now also return the name of the PSF used in the fit
(additional to the fit parameters) in the output catalog, an important thing to check or to
know after the fits have been done.

Note:
For now, NO INTERPOLATION between the PSFs is done.
This might be another mode in the future, for now all it does is CHOOSE a pre-given PSF!