prepare_galfit should work fine now. Therefore gala_bridge should work fine as well. Problem: GALFIT returns SEGMENTATION FAULT most of the times.
GALA internal constraints (e.g. re ) currently work on ALL BANDS
and due to GALFIT output format and the internal fittab table,
Galapagos will now NOT be compatible with previous versions of
GALFIT. GALA now NEEDS the fits table as an output even on 1-band
data, otherwise values will not be filled up with good values
that can be used for deblending.
The output format to hand over to GALFIT is still compatible
with old versions, but we should remove this from GALA
(as well as the version input in general, all it was needed for
was the format of the input file).

Rest should work fine.
To do: Section with batch_list