introduced new queueing system into galapagos.
It runs in the test script (not uploaded), I'm currently testing the test run on 9-band real GAMA data.

The idea is the following:
As before, galapagos check whether the next brightest object is closer than a certain user-specified distance to a running object. If so, the old script stopped and waited. It now blocks this object and proceedes with the next brightest if this is far enough away.

On aditional feature:
galapagos also does NOT fit objects that are closer than (currently) 0.5* above specified distance. This way, we can avoid fitting a faintish object which is JUST outsode the blocking radius but very close to a bright (and blocked object) which is JUST inside. I used the same distance initially, but that very quickly lead to the whole survey area being blocked because it is an iterative process.
The new scheme is faster while still securing a useful system.