Had a look at the speed of the bridge and the sky determination, after consultation with Marco
removed some resistant_means, simplified the system (while making sure that at least on GAMA data, the result is the same. There was a 3 step process, now it's only 2 step, where the starting values for the last step are less accurate, but much faster.
the old scheme is still in, only commented, so it's easy to go back.
most changes in gala_bridge.pro

moved the creation of a needed array to the bridge in order to only do it 1 time and not for each band individually. saves another few seconds per band

replaced ALL fits_read with readfits to save CPU time. readfits is twice as fast (but might be faulty at some point, the different procedures have different strangths and weaknesses)

Other than this, the actual RING-system applied by galapagos only takes 0.4 of a second per band, so speeding this up by a trained system would not improve things much. Skipping this for now.

Got the typical CPU time for setting up one objects from ~150 seconds to ~80 seconds, so speedup on this part nearly 50%.