fixed a bug in the new queue.
the queue (as previously) only checke todo[0] whether it has arleady been done. If this was a blocked object, it SHOULD have checked whether todo[1] has been done previously, but instead started the setup and the fit of it, making basically every object twice over a secondary run.
This is now fixed and object numbers jump around and block each other as they should. I'm optimistic that that was it.