reordered parts of gala_bridge, so it's more readable
adapted read_sersic_result and update_table to be able to read in B/D results
I have adapted the routines we had (read_sersic_results) rather than creating a new one, but working with flags. The routines are now MUCH longer, and I did not test them yet (the fitting bit does not work, so the real output files do not exist, so reading them in would create errors in places), but they SHOULD work, these routines are not very complicated.