major bug fix!!!
It happened that galapagos introduced a neoghbor VERY far off the postage stamp.
This came down to some string comparison between contrib_target_frames and the tile names in the SExtractor catalog (e.g. naming convention).
I have changed contrib_target to return the filename apparent in the sextractor catalogue instead of the one in the band (index 0 instead of 1) and it now seems to work.