Galapagos now takes the expsosure time for the SExtractor images from the setup file, rather than from the image header. While this seems more complicated, it is more in line with the other filters and allows more flexibility for different surveys (we encountered that some software needs the header keyword to be different, which is avoided this way.)
BUG FIX & NEW INPUT:
The input file now requires a line E25), which is the minimum number of 'good' bands (bands with data) in order for the fit to actually be carried out.
In previous versions, it could happen that Galapagos restricts the maximum number of degrees of freedom to 0.
This would mean that the galfit fit is being carried out, but without free parameters.
This would write out a galfit 'output' file, with all the model parameters being the input parameters, mostly as derived by SExtractor.
If this object the becomes a secondary for a different object, Galapagos would hold its parameters fixed at these values, which is obviously a very bad idea.
Not carrying out the 'fit' will prevent this, the object would be implemented as an object with free parameters.
However, such cases should BEST be dealt with in the sextractor mask/weight, so this problem would not even appear.
This feature is a security measure to prevent the worst.
These objects can be identified in the Galapagos output catalogue by having flag_galfit (or flag_galfit_bd) == -1